refactoring, fixed terminal cooked  vs raw mode bugs, simplified main.rs, fixed bug allowing repeated keypresses during timer to increase the time logged

+++ src/util/types.rs
@@ -1,6 +1,6 @@
 #[derive(Debug, serde::Serialize, serde::Deserialize)]
 pub struct TimeLog {
     pub id: u32,
-    pub time_spent: [i8;3],
+    pub time_spent: [u64;3],
     pub date: String,
 }
